Separate defconfig creation from re-use logic
At the moment when _use_current=y is passed to the package
compilation - the script still re-creates the defconfig which can
cause errors and cause the menuconfig to show up outta nowhere.

Fix this by bypassing the logic completely, I think that
anyone using the _use_current=y knows what they're doing
so we should be fine.